escaner/requirements.txt
flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
jinja2==3.1.2
flask-cors==4.0.0
fpdf2==2.7.8